Compartmentalize audio functions/files.
